requirements.txt
torch
torchvision
ignite
Pillow
captcha
pycrypto
tqdm